lints and mypy
